# public/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Life@USTC API",
    "version": "1.0.0",
    "description": "OpenAPI document generated from Next.js routes by next-openapi-gen"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/api/comments": {
      "get": {
        "operationId": "get-api-comments",
        "summary": "List comments for a target.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "query",
            "name": "targetType",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "targetId",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "sectionId",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "teacherId",
            "schema": {},
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/commentsListResponseSchema"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-api-comments",
        "summary": "Create one comment.",
        "description": "",
        "tags": ["Api"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/commentCreateRequestSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/idResponseSchema"
                }
              }
            }
          }
        }
      }
    },
    "/api/courses": {
      "get": {
        "operationId": "get-api-courses",
        "summary": "List courses with search and pagination.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {},
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/paginatedCourseResponseSchema"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/homeworks/{id}/completion": {
      "put": {
        "operationId": "put-api-homeworks-{id}-completion",
        "summary": "Update completion state for one homework.",
        "description": "",
        "tags": ["Api"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "example": "123"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/homeworkCompletionRequestSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/homeworkCompletionResponseSchema"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "commentsListResponseSchema": {
        "type": "object",
        "properties": {
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/commentNodeSchema"
            }
          },
          "hiddenCount": {
            "type": "integer",
            "minimum": 0
          },
          "viewer": {
            "$ref": "#/components/schemas/viewerContextSchema"
          },
          "target": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "targetId": {
                "oneOf": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "sectionId": {
                "type": "integer",
                "nullable": true
              },
              "teacherId": {
                "type": "integer",
                "nullable": true
              },
              "sectionTeacherId": {
                "type": "integer",
                "nullable": true
              },
              "homeworkId": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "type",
              "targetId",
              "sectionId",
              "teacherId",
              "sectionTeacherId",
              "homeworkId"
            ]
          }
        },
        "required": ["comments", "hiddenCount", "viewer", "target"]
      },
      "commentCreateRequestSchema": {
        "type": "object",
        "properties": {
          "targetType": {
            "$ref": "#/components/schemas/commentTargetTypeSchema"
          },
          "targetId": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "sectionId": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "teacherId": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8000
          },
          "visibility": {
            "allOf": [
              {
                "$ref": "#/components/schemas/commentVisibilitySchema"
              }
            ]
          },
          "isAnonymous": {
            "type": "boolean"
          },
          "parentId": {
            "type": "string",
            "nullable": true
          },
          "attachmentIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["targetType", "body"]
      },
      "idResponseSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": ["id"]
      },
...
      "paginatedCourseResponseSchema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/courseSchema"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/pageMetaSchema"
          }
        },
        "required": ["data", "pagination"]
      },
...
      "homeworkCompletionRequestSchema": {
        "type": "object",
        "properties": {
          "completed": {
            "type": "boolean"
          }
        },
        "required": ["completed"]
      },
      "homeworkCompletionResponseSchema": {
        "type": "object",
        "properties": {
          "completed": {
            "type": "boolean"
          },
          "completedAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/dateTimeSchema"
              }
            ]
          }
        },
        "required": ["completed", "completedAt"]
      },
...
      "commentVisibilitySchema": {
        "type": "string",
        "enum": ["public", "logged_in_only", "anonymous"]
      },
      "commentTargetTypeSchema": {
        "type": "string",
        "enum": ["section", "course", "teacher", "section-teacher", "homework"]
      },
...
      "dateTimeSchema": {
        "type": "string"
      },
      "pageMetaSchema": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        },
        "required": ["page", "pageSize", "total", "totalPages"]
      },
...
      "commentNodeSchema": {
        "type": "string"
      },
      "viewerContextSchema": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "isAdmin": {
            "type": "boolean"
          },
          "isAuthenticated": {
            "type": "boolean"
          },
          "isSuspended": {
            "type": "boolean"
          },
          "suspensionReason": {
            "type": "string",
            "nullable": true
          },
          "suspensionExpiresAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/dateTimeSchema"
              }
            ]
          }
        },
        "required": [
          "userId",
          "name",
          "image",
          "isAdmin",
          "isAuthenticated",
          "isSuspended",
          "suspensionReason",
          "suspensionExpiresAt"
        ]
      },
...
      "courseSchema": {
        "type": "object",
        "properties": {
          "category": {
            "type": "object"
          },
          "classType": {
            "type": "object"
          },
          "classify": {
            "type": "object"
          },
          "educationLevel": {
            "type": "object"
          },
          "gradation": {
            "type": "object"
          },
          "type": {
            "type": "object"
          }
        },
        "required": [
          "category",
          "classType",
          "classify",
          "educationLevel",
          "gradation",
          "type"
        ]
      },
...
    },
...
  }
}